Refactored listitem handling
