+++ Commands/Lists/AddListItem.cs
@@ -9,6 +9,7 @@ using OfficeDevPnP.Core.Utilities;
 using SharePointPnP.PowerShell.CmdletHelpAttributes;
 using SharePointPnP.PowerShell.Commands.Base.PipeBinds;
 using SharePointPnP.PowerShell.Commands.Taxonomy;
+using SharePointPnP.PowerShell.Commands.Utilities;
 
 // IMPORTANT: If you make changes to this cmdlet, also make the similar/same changes to the Set-PnPListItem Cmdlet
 
@@ -119,173 +120,15 @@ namespace SharePointPnP.PowerShell.Commands.Lists
 
                 if (Values != null)
                 {
-                    Hashtable values = Values ?? new Hashtable();
-                    // Load all list fields and their types
-                    var fields = ClientContext.LoadQuery(list.Fields.Include(f => f.Id, f => f.InternalName, f => f.Title, f => f.TypeAsString));
-                    ClientContext.ExecuteQueryRetry();
-
-                    foreach (var key in values.Keys)
-                    {
-                        var field = fields.FirstOrDefault(f => f.InternalName == key as string || f.Title == key as string);
-                        if (field != null)
+                    item = ListItemHelper.UpdateListItem(item, Values, false,
+                        (warning) =>
                         {
-                            switch (field.TypeAsString)
-                            {
-                                case "User":
-                                case "UserMulti":
-                                    {
-                                        var userValues = new List<FieldUserValue>();
-
-                                        var value = values[key];
-                                        if (value.GetType().IsArray)
-                                        {
-                                            foreach (var arrayItem in value as object[])
-                                            {
-                                                int userId;
-                                                if (!int.TryParse(arrayItem as string, out userId))
-                                                {
-                                                    var user = SelectedWeb.EnsureUser(arrayItem as string);
-                                                    ClientContext.Load(user);
-                                                    ClientContext.ExecuteQueryRetry();
-                                                    userValues.Add(new FieldUserValue() { LookupId = user.Id });
-                                                }
-                                                else
-                                                {
-                                                    userValues.Add(new FieldUserValue() { LookupId = userId });
-                                                }
-                                            }
-                                            item[key as string] = userValues.ToArray();
-                                        }
-                                        else
-                                        {
-                                            int userId;
-                                            if (!int.TryParse(value as string, out userId))
-                                            {
-                                                var user = SelectedWeb.EnsureUser(value as string);
-                                                ClientContext.Load(user);
-                                                ClientContext.ExecuteQueryRetry();
-                                                item[key as string] = new FieldUserValue() { LookupId = user.Id };
-                                            }
-                                            else
-                                            {
-                                                item[key as string] = new FieldUserValue() { LookupId = userId };
-                                            }
-                                        }
-                                        item.Update();
-                                        break;
-                                    }
-                                case "TaxonomyFieldType":
-                                case "TaxonomyFieldTypeMulti":
-                                    {
-                                        var value = values[key];
-                                        if (value.GetType().IsArray)
-                                        {
-                                            var taxSession = ClientContext.Site.GetTaxonomySession();
-                                            var terms = new List<KeyValuePair<Guid, string>>();
-                                            foreach (var arrayItem in value as object[])
-                                            {
-                                                TaxonomyItem taxonomyItem;
-                                                Guid termGuid = Guid.Empty;
-                                                if (!Guid.TryParse(arrayItem as string, out termGuid))
-                                                {
-                                                    // Assume it's a TermPath
-                                                    taxonomyItem = ClientContext.Site.GetTaxonomyItemByPath(arrayItem as string);
-                                                }
-                                                else
-                                                {
-                                                    taxonomyItem = taxSession.GetTerm(termGuid);
-                                                    ClientContext.Load(taxonomyItem);
-                                                    ClientContext.ExecuteQueryRetry();
-                                                }
-
-
-
-                                                terms.Add(new KeyValuePair<Guid, string>(taxonomyItem.Id, taxonomyItem.Name));
-                                            }
-
-                                            TaxonomyField taxField = ClientContext.CastTo<TaxonomyField>(field);
-
-                                            taxField.EnsureProperty(tf => tf.AllowMultipleValues);
-
-                                            if (taxField.AllowMultipleValues)
-                                            {
-                                                var termValuesString = String.Empty;
-                                                foreach (var term in terms)
-                                                {
-                                                    termValuesString += "-1;#" + term.Value + "|" + term.Key.ToString("D") + ";#";
-                                                }
-
-                                                termValuesString = termValuesString.Substring(0, termValuesString.Length - 2);
-
-                                                var newTaxFieldValue = new TaxonomyFieldValueCollection(ClientContext, termValuesString, taxField);
-                                                taxField.SetFieldValueByValueCollection(item, newTaxFieldValue);
-                                                item.Update();
-                                                ClientContext.ExecuteQueryRetry();
-                                            }
-                                            else
-                                            {
-                                                WriteWarning($@"You are trying to set multiple values in a single value field. Skipping values for field ""{field.InternalName}""");
-                                            }
-                                        }
-                                        else
-                                        {
-                                            Guid termGuid = Guid.Empty;
-                                            if (!Guid.TryParse(value as string, out termGuid))
-                                            {
-                                                // Assume it's a TermPath
-                                                var taxonomyItem = ClientContext.Site.GetTaxonomyItemByPath(value as string);
-                                                termGuid = taxonomyItem.Id;
-                                            }
-                                            item[key as string] = termGuid.ToString();
-                                        }
-                                        item.Update();
-                                        break;
-                                    }
-                                case "Lookup":
-                                case "LookupMulti":
-                                    {
-                                        int[] multiValue;
-                                        if (values[key] is Array)
-                                        {
-                                            var arr = (object[])values[key];
-                                            multiValue = new int[arr.Length];
-                                            for (int i = 0; i < arr.Length; i++)
-                                            {
-                                                multiValue[i] = int.Parse(arr[i].ToString());
-                                            }
-                                        }
-                                        else
-                                        {
-                                            string valStr = values[key].ToString();
-                                            multiValue = valStr.Split(',', ';').Select(int.Parse).ToArray();
-                                        }
-
-                                        var newVals = multiValue.Select(id => new FieldLookupValue { LookupId = id }).ToArray();
-
-                                        FieldLookup lookupField = ClientContext.CastTo<FieldLookup>(field);
-                                        lookupField.EnsureProperty(lf => lf.AllowMultipleValues);
-                                        if (!lookupField.AllowMultipleValues && newVals.Length > 1)
-                                        {
-                                            WriteWarning($@"You are trying to set multiple values in a single value field. Skipping values for field ""{field.InternalName}""");
-                                        }
-
-                                        item[key as string] = newVals;
-                                        item.Update();
-                                        break;
-                                    }
-                                default:
-                                    {
-                                        item[key as string] = values[key];
-                                        item.Update();
-                                        break;
-                                    }
-                            }
-                        }
-                        else
+                            WriteWarning(warning);
+                        },
+                        (terminatingErrorMessage, terminatingErrorCode) =>
                         {
-                            ThrowTerminatingError(new ErrorRecord(new Exception("Field not present in list"), "FIELDNOTINLIST", ErrorCategory.InvalidData, key));
-                        }
-                    }
+                            ThrowTerminatingError(new ErrorRecord(new Exception(terminatingErrorMessage), terminatingErrorCode, ErrorCategory.InvalidData, this));
+                        });
                 }
 
                 item.Update();
