+++ Commands/SharePointPnP.PowerShell.Commands.csproj
@@ -654,6 +654,7 @@
     <Compile Include="UserProfiles\SetUserProfileProperty.cs" />
     <Compile Include="Utilities\DynamicLinq.cs" />
     <Compile Include="Utilities\HttpHelper.cs" />
+    <Compile Include="Utilities\ListItemHelper.cs" />
     <Compile Include="Webhooks\RemoveWebhookSubscription.cs" />
     <Compile Include="Webhooks\SetWebhookSubscription.cs" />
     <Compile Include="Webhooks\GetWebhookSubscriptions.cs" />
